extract: cancel modal behavior
Modal cancelling behavior has been extracted
out of the template base into its own hook.
